// src/app/interfaces/pokemon.interfaces.ts
export interface Pokemon {
    id:string;
    name:string;
    pic:string;
    favorite:boolean;
}